Support eslint's no-useless-escape rule
This will make sure to support eslint's no-useless-escape rule.

"Useless escapes in the output mean useless escapes in the input,
though!"

On one hand, yes. On the other hand, Regex101 doesn't actually have an
XRegExp setting, so you have to use PCRE x-flagged regexes, which
differ from JavaScript in several fortunately-minor ways, but one of
them is that some escapes that are useless in JavaScript are required
in PCRE.

(Specifically, PCRE requires `[\/]` but `[/]` is enough in JS.)